requirements.txt
pandas==0.24.1
Pillow==8.2.0
